Move tests inside subfolders
https://coveord.atlassian.net/browse/JSUI-2373
